삭제, 수정 수정
수정수정

+++ routes/posts.js
@@ -112,6 +112,7 @@ router.put('/:postId', async (req, res) => {
     const { postId } = req.params;
     const { title, content } = req.body;
     const { Authorization } = req.cookies;
+    const post = Post.findOne({ where: postId });
 
     if (!req.body) {
       return res.status(412).json({ errorMessage: '데이터 형식이 올바르지 않습니다.' });
@@ -128,7 +129,9 @@ router.put('/:postId', async (req, res) => {
     if (!validateToken(Authorization)) {
       return res.status(403).json({ errorMessage: '전달된 쿠키에서 오류가 발생하였습니다.' });
     }
-
+    if (getTokenPayload !== post.userId) {
+      return res.status(401).json({ errorMessage: '게시글이 정상적으로 수정되지 않았습니다.' });
+    }
     Post.update(
       {
         title, content,
@@ -144,11 +147,13 @@ router.put('/:postId', async (req, res) => {
   return res.status(777).send('도대체 어떻게 온거야.');
 });
 // 게시글 삭제
+// eslint-disable-next-line consistent-return
 router.delete('/:postId', async (req, res) => {
   try {
     const { postId } = req.params;
     const { Authorization } = req.cookies;
-    if (!postId) {
+    const post = Post.findOne({ where: postId });
+    if (!post || !postId) {
       return res.status(404).json({ errorMessage: '게시글이 존재하지 않습니다.' });
     }
     if (!parameterVerification(postId)) {
@@ -159,6 +164,9 @@ router.delete('/:postId', async (req, res) => {
     }
     if (!validateToken(Authorization)) {
       return res.status(403).json({ errorMessage: '전달된 쿠키에서 오류가 발생하였습니다.' });
+    }
+    if (getTokenPayload !== post.userId) {
+      return res.status(401).json({ errorMessage: '게시글이 정상적으로 삭제되지 않았습니다.' });
     }
     Post.destroy({
       where: { postId },
@@ -172,3 +180,4 @@ router.delete('/:postId', async (req, res) => {
 module.exports = router;
 
 // 게시글 없어도 삭제가 계속됨.
+// 삭제 권한 한정
